Move ProgressReporter to UI.Progress
