+++ src/KappaDuck.Quack/UI/Progress/ProgressReporter.cs
@@ -1,7 +1,7 @@
 // Copyright (c) KappaDuck.
 // Licensed under the MIT license.
 
-namespace KappaDuck.Quack.UI.Progress.Reporters;
+namespace KappaDuck.Quack.UI.Progress;
 
 /// <summary>
 /// Synchronous determinate progress reporter handed to <see cref="ProgressBarBase.Start(Action{ProgressReporter}, int)"/>.
