+++ tests/Unit.Tests/UI/Progress/ProgressReporterTests.cs
@@ -2,9 +2,8 @@
 // Licensed under the MIT license.
 
 using KappaDuck.Quack.UI.Progress;
-using KappaDuck.Quack.UI.Progress.Reporters;
 
-namespace Unit.Tests.UI.Progress.Reporters;
+namespace Unit.Tests.UI.Progress;
 
 internal sealed class ProgressReporterTests
 {
